// types/movieById.ts
export type MoviesByIdType = {
  Title: string;
  Year: string;
  Rated: string;
  Released: string;
  Genre: string;
  Director: string;
  Writer: string;
  Actors: string;
  Plot: string;
  Language: string;
  Country: string;
  Poster: string;
  imdbRating: string;
  Type: string;
  Runtime: string;
  imdbVotes: string;
};